Implement a post processor to resolve remote image links
This commit demonstrates the capabilities of the post
processing engine by introducing a remote image resolver,
which is able to insert images into a document.
This resolver needs to be passed a list of accepted urls for
security reasons (this is still an area of work).
The way this resolver works is by going through the document
and whenever it encounters a placeholder it extracts the
link, checks whether it is valid and points to an allowed
domain, downloads the image and replaces the paragraph with
it.

Signed-off-by: Anton Oellerer <a.oellerer@docu-tools.com>